#177: Renamed services package to rosie
