+++ src/main/java/io/codiga/plugins/jetbrains/rosie/CodigaConfigSchemaProviderFactory.java
@@ -1,4 +1,4 @@
-package io.codiga.plugins.jetbrains.services;
+package io.codiga.plugins.jetbrains.rosie;
 
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.project.Project;
