+++ src/main/java/io/codiga/plugins/jetbrains/starter/RosieRulesCacheUpdateHandler.java
@@ -1,8 +1,8 @@
 package io.codiga.plugins.jetbrains.starter;
 
 import static io.codiga.plugins.jetbrains.Constants.LOGGER_NAME;
-import static io.codiga.plugins.jetbrains.services.CodigaConfigFileUtil.collectRulesetNames;
-import static io.codiga.plugins.jetbrains.services.CodigaConfigFileUtil.findCodigaConfigFile;
+import static io.codiga.plugins.jetbrains.rosie.CodigaConfigFileUtil.collectRulesetNames;
+import static io.codiga.plugins.jetbrains.rosie.CodigaConfigFileUtil.findCodigaConfigFile;
 
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
