+++ src/main/resources/META-INF/yaml-features.xml
@@ -1,11 +1,11 @@
 <idea-plugin>
     <extensions defaultExtensionNs="com.intellij">
         <applicationService
-                serviceInterface="io.codiga.plugins.jetbrains.services.Rosie"
-                serviceImplementation="io.codiga.plugins.jetbrains.services.RosieImpl"
-                testServiceImplementation="io.codiga.plugins.jetbrains.services.RosieTest"
+                serviceInterface="io.codiga.plugins.jetbrains.rosie.Rosie"
+                serviceImplementation="io.codiga.plugins.jetbrains.rosie.RosieImpl"
+                testServiceImplementation="io.codiga.plugins.jetbrains.rosie.RosieTest"
         />
-        <postStartupActivity implementation="io.codiga.plugins.jetbrains.starter.RosieRulesCacheUpdater"/>
+        <postStartupActivity implementation="io.codiga.plugins.jetbrains.starter.RosieStartupActivity"/>
         <externalAnnotator
                 language=""
                 implementationClass="io.codiga.plugins.jetbrains.annotators.RosieAnnotator"/>
@@ -15,6 +15,6 @@
     </extensions>
 
     <extensions defaultExtensionNs="JavaScript.JsonSchema">
-        <ProviderFactory implementation="io.codiga.plugins.jetbrains.services.CodigaConfigSchemaProviderFactory"/>
+        <ProviderFactory implementation="io.codiga.plugins.jetbrains.rosie.CodigaConfigSchemaProviderFactory"/>
     </extensions>
 </idea-plugin>
